+++ ShoppingCart.Infrastructure/Interfaces/IUserMapper.cs
@@ -4,7 +4,7 @@ namespace ShoppingCart.Infrastructure.Interfaces;
 
 public interface IUserMapper
 {
-    Task<User?> GetUserByDniAsync(double dni);
-    Task<bool> ExistsAsync(double dni);
+    Task<User?> GetUserByDniAsync(long dni);
+    Task<bool> ExistsAsync(long dni);
     Task<User?> GetUserByIdAsync(int userId);
 }
